Move tree representation implementation to new provider LibraryRepresentationActivityItemProvider
